+++ lib/kdtree.cpp
@@ -59,7 +59,7 @@ KDTree::precompute(EigenPointCloudRef querypoints,
 }
 
 std::size_t
-KDTree::radius_search(const double* query,
+KDTree::radius_search(const float* query,
                       double radius,
                       RadiusSearchResult& result) const
 {
@@ -70,7 +70,7 @@ KDTree::radius_search(const double* query,
 }
 
 std::size_t
-KDTree::radius_search_with_distances(const double* query,
+KDTree::radius_search_with_distances(const float* query,
                                      double radius,
                                      RadiusSearchDistanceResult& result) const
 {
